refactor blockfusion
